TSK-967: Move DbCleaner to taskana-data
