--- lib/taskana-core/src/test/resources/sql/drop-tables.sql
@@ -1,11 +1,0 @@
-DROP TABLE TASKANA_SCHEMA_VERSION;
-DROP TABLE ATTACHMENT;
-DROP TABLE TASK;
-DROP TABLE WORKBASKET_ACCESS_LIST;
-DROP TABLE WORKBASKET;
-DROP TABLE DISTRIBUTION_TARGETS;
-DROP TABLE CLASSIFICATION;
-DROP TABLE OBJECT_REFERENCE;
-DROP TABLE SCHEDULED_JOB;
-DROP SEQUENCE SCHEDULED_JOB_SEQ;
-COMMIT;
