+++ lib/taskana-data/src/main/resources/sql/clear/clear-db.sql
