+++ lib/taskana-data/src/main/resources/sql/clear/clear-history-events.sql
