+++ lib/taskana-data/src/main/resources/sql/clear/drop-tables.sql
@@ -7,6 +7,6 @@ DROP TABLE DISTRIBUTION_TARGETS;
 DROP TABLE CLASSIFICATION;
 DROP TABLE OBJECT_REFERENCE;
 DROP TABLE SCHEDULED_JOB;
-DROP TABLE HISTORY_EVENTS;
+DROP TABLE HISTORY_EVENTS IF EXISTS;
 DROP SEQUENCE SCHEDULED_JOB_SEQ;
 COMMIT;
