--- lib/taskana-data/src/main/resources/sql/sample-data/clear-db.sql
@@ -1,11 +1,0 @@
--- the order is important!
-DELETE FROM ATTACHMENT;
-DELETE FROM TASK;
-DELETE FROM WORKBASKET_ACCESS_LIST;
-DELETE FROM WORKBASKET;
-DELETE FROM DISTRIBUTION_TARGETS;
-DELETE FROM CLASSIFICATION;
-DELETE FROM OBJECT_REFERENCE;
--- do not clean JOB table
--- DELETE FROM SCHEDULED_JOB;
-COMMIT;
